Finish javascript code -------------------------------------------
I wrote the code in a way that leaves a place for future improvements.
In terms of flexibility, I returned the results as strings instead of
logging them using "console.log()" so that they can be easily used on
a webpage in the future. As for encapsulation, I put the game-related
functions and variables in a class since it's good a practice to
organize the code.